+++ govtool/frontend/src/components/molecules/types.ts
@@ -1,0 +1,7 @@
+export type AutomatedVotingCardProps = {
+  description: string;
+  onClickDelegate: () => void;
+  onClickInfo: () => void;
+  title: string;
+  votingPower: string | number;
+};